+++ src/Adapter/Combination/Mandatory.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("../Subject/Subject.mjs").Subject} Subject */
+
+/**
+ * @typedef {Subject} Mandatory
+ */