+++ src/Adapter/Get/GetResult.mjs
@@ -1,5 +1,5 @@
 /** @typedef {import("../Page/PAGE.mjs").PAGE} PAGE */
 
 /**
- * @typedef {{page: PAGE, data: {[key: string]: *}, can_back?: boolean}} GetResult
+ * @typedef {{page: PAGE, data: {[key: string]: *}, can_back: boolean}} GetResult
  */