+++ src/Adapter/IntendedDegreeProgram2/IntendedDegreeProgram2Element.mjs
@@ -1,5 +1,6 @@
 import { ELEMENT_TAG_NAME_PREFIX } from "../Element/ELEMENT_TAG_NAME_PREFIX.mjs";
 import { FormElement } from "../Form/FormElement.mjs";
+import { MandatoryElement } from "../Mandatory/MandatoryElement.mjs";
 import { PAGE_INTENDED_DEGREE_PROGRAM_2 } from "../Page/PAGE.mjs";
 import { TitleElement } from "../Title/TitleElement.mjs";
 
@@ -95,6 +96,26 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
 
         this.#chosen_intended_degree_program_2_function(
             {
+                "single-choice": this.#intended_degree_program_2.combination["single-choice"] !== null ? Object.fromEntries(Object.entries(this.#form_element.getStartsWithInputs(
+                    "single-choice_"
+                )).map(([
+                    id,
+                    input_element
+                ]) => [
+                        id,
+                        input_element.value
+                    ])) : null,
+                "multiple-choice": this.#intended_degree_program_2.combination["multiple-choice"] !== null ? Object.fromEntries(Object.entries(this.#form_element.getStartsWithInputs(
+                    "multiple-choice_"
+                )).map(([
+                    id,
+                    input_element
+                ]) => [
+                        id,
+                        [
+                            ...input_element.selectedOptions
+                        ].map(option_element => option_element.value)
+                    ])) : null,
                 "further-information": this.#form_element.inputs["further-information"].value.replaceAll("\r\n", "\n").replaceAll("\r", "\n")
             }
         );
@@ -110,7 +131,51 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
         ));
 
         this.#form_element = FormElement.new(
-            this.#css_api
+            this.#css_api,
+            () => {
+                if (this.#intended_degree_program_2.combination["multiple-choice"] === null) {
+                    return true;
+                }
+
+                for (const multiple_choice of this.#intended_degree_program_2.combination["multiple-choice"]) {
+                    const input_element = this.#form_element.inputs[`multiple-choice_${multiple_choice.id}`];
+
+                    const values = [
+                        ...input_element.selectedOptions
+                    ].map(option_element => option_element.value);
+
+                    let ect = 0;
+                    for (const choice of multiple_choice.choices) {
+                        if (values.includes(choice.id)) {
+                            ect += choice.ect;
+                        }
+                    }
+
+                    if (ect < multiple_choice.ect) {
+                        this.#form_element.setCustomValidationMessage(
+                            input_element,
+                            `Please select more${this.#label_service.getEctLabel(
+                                multiple_choice.ect - ect,
+                                false
+                            )}!`
+                        );
+                        return false;
+                    }
+
+                    if (ect > multiple_choice.ect) {
+                        this.#form_element.setCustomValidationMessage(
+                            input_element,
+                            `Please select less${this.#label_service.getEctLabel(
+                                ect - multiple_choice.ect,
+                                false
+                            )}!`
+                        );
+                        return false;
+                    }
+                }
+
+                return true;
+            }
         );
 
         this.#form_element.addTitle(
@@ -141,12 +206,66 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
             this.#intended_degree_program_2.combination
         );
 
-        this.#form_element.addSubtitle(
-            "If you have already studied the choosen subject and would like to continue in an advanced semester, please enter it here:"
-        );
+        if (this.#intended_degree_program_2.combination["single-choice"] !== null || this.#intended_degree_program_2.combination["multiple-choice"] !== null) {
+            this.#form_element.addSubtitle(
+                "Please choose your combination of subjects from the given selection below"
+            );
 
-        const further_information_element = this.#form_element.addTextarea(
-            "further-information"
+            if (this.#intended_degree_program_2.combination["single-choice"] !== null) {
+                for (const single_choice of this.#intended_degree_program_2.combination["single-choice"]) {
+                    const single_choice_element = this.#form_element.addInput(
+                        this.#label_service.getSingleChoiceLabel(
+                            single_choice
+                        ),
+                        "select",
+                        `single-choice_${single_choice.id}`,
+                        true
+                    );
+                    single_choice_element.required = true;
+
+                    for (const choice of single_choice.choices) {
+                        const option_element = document.createElement("option");
+                        option_element.text = this.#label_service.getChoiceLabel(
+                            choice
+                        );
+                        option_element.value = choice.id;
+                        single_choice_element.appendChild(option_element);
+                    }
+                }
+            }
+
+            if (this.#intended_degree_program_2.combination["multiple-choice"] !== null) {
+                for (const multiple_choice of this.#intended_degree_program_2.combination["multiple-choice"]) {
+                    const multiple_choice_element = this.#form_element.addInput(
+                        this.#label_service.getMultipleChoiceLabel(
+                            multiple_choice
+                        ),
+                        "select",
+                        `multiple-choice_${multiple_choice.id}`,
+                        true
+                    );
+                    multiple_choice_element.innerHTML = "";
+                    multiple_choice_element.multiple = true;
+                    multiple_choice_element.required = true;
+                    multiple_choice_element.size = multiple_choice.choices.length;
+
+                    for (const choice of multiple_choice.choices) {
+                        const option_element = document.createElement("option");
+                        option_element.text = this.#label_service.getChoiceLabel(
+                            choice
+                        );
+                        option_element.value = choice.id;
+                        multiple_choice_element.appendChild(option_element);
+                    }
+                }
+            }
+        }
+
+        const further_information_element = this.#form_element.addInput(
+            "If you have already studied the choosen subject and would like to continue in an advanced semester, please enter it here",
+            "textarea",
+            "further-information",
+            true
         );
 
         this.#form_element.addButtons(
@@ -158,7 +277,41 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
 
         this.#shadow.appendChild(this.#form_element);
 
+        this.#shadow.appendChild(MandatoryElement.new(
+            this.#css_api
+        ));
+
         if (this.#intended_degree_program_2.values !== null) {
+            if (this.#intended_degree_program_2.values["single-choice"] !== null) {
+                for (const [
+                    id,
+                    input_element
+                ] of Object.entries(this.#form_element.getStartsWithInputs(
+                    "single-choice_"
+                ))) {
+                    if (id in this.#intended_degree_program_2.values["single-choice"]) {
+                        input_element.value = this.#intended_degree_program_2.values["single-choice"][id];
+                    }
+                }
+            }
+
+            if (this.#intended_degree_program_2.values["multiple-choice"] !== null) {
+                for (const [
+                    id,
+                    input_element
+                ] of Object.entries(this.#form_element.getStartsWithInputs(
+                    "multiple-choice_"
+                ))) {
+                    if (id in this.#intended_degree_program_2.values["multiple-choice"]) {
+                        for (const option_element of input_element.options) {
+                            if (this.#intended_degree_program_2.values["multiple-choice"][id].includes(option_element.value)) {
+                                option_element.selected = true;
+                            }
+                        }
+                    }
+                }
+            }
+
             further_information_element.value = this.#intended_degree_program_2.values["further-information"];
         }
     }