+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2022-10-04-1-alpha
+
+Changes:
+
+- Single and multiple choice
+
 ## v2022-10-03-1-alpha
 
 Changes: